#50 Show in list real files instead of fake

+++ src/Zerebral/BusinessBundle/Model/Assignment/map/AssignmentTableMap.php
@@ -48,7 +48,7 @@ class AssignmentTableMap extends TableMap
         $this->addForeignKey('assignment_category_id', 'AssignmentCategoryId', 'TINYINT', 'assignment_categories', 'id', true, 2, null);
         $this->addColumn('name', 'Name', 'VARCHAR', true, 200, null);
         $this->addColumn('description', 'Description', 'LONGVARCHAR', false, null, null);
-        $this->addColumn('max_points', 'MaxPoints', 'TINYINT', true, 2, null);
+        $this->addColumn('max_points', 'MaxPoints', 'SMALLINT', true, 2, null);
         $this->addColumn('due_at', 'DueAt', 'TIMESTAMP', false, null, null);
         // validators
         $this->addValidator('id', 'required', 'propel.validator.RequiredValidator', '', 'The field id is required.');
